moving site-utils from site/component to site/hook
